Fixed Ep error when reviewing PEB
Reviewing BMA after BMC using spm_dcm_peb_review would throw an error if
the GCM structure provided did not contain the field 'Ep'. This would be
the case when the models used for BMC would be defined on DCM
specifications which have not been fit.

For example, when template DCMs are used to define a full model, and
subsequent reduced models are derived from it to construct a model space
(encoded as columns of the GCM array), these will not contain any priors
or posteriors. This model space can be constructed manually or using the
batch interface, and in either case the model space will not have the Ep
field.

The error shows up during review when the 'View matrix' option is
selected, which shows the parameter estimates in a matrix instead of
barplots. For converting the vector of estimates to the dimensions of
the matrix, the exact dimensions are fetched from the size of the
relevant field in Ep of the first DCM. This leads to an error if that
DCM is merely a template specification and not a fitted model, therefore
does not contain Ep.

This is the error message:
    Reference to non-existent field 'Ep'.

    Error in spm_dcm_peb_review>update_view (line 302)
        [i,j,k] = size(eval(['DCM{1}.Ep.' sel_field]));

    Error in spm_dcm_peb_review>selected_threshold_changed (line 715)
    update_view();

    Error while evaluating UIControl Callback.

The current fix is to pick the relevant field from the top-level field
structure, so for the B{1} matrix, instead of DCM.Ep.B{1}, this will
check the dimensions of DCM.B{1}. This may not work for fields which are
not the A/B/C matrices (eg. G or other neuronal model parameters). This
may not work for fMRI either because the matrices are not inside cell
arrays and are 3-dimensional instead (confirm?).